Intial GCC testsuite support.

+++ update-gcc.sh
@@ -39,12 +39,24 @@ if test -e "$d_gccsrc/gcc/d" -o -e "$d_gccsrc/libphobos"; then
     exit 1
 fi
 
+d_test=$d_gccsrc/gcc/testsuite
+# remove testsuite sources
+test -d "$d_test/gdc.test" && rm -r "$d_test/gdc.test"
+test -e "$d_test/lib/gdc.exp" && rm "$d_test/lib/gdc.exp"
+test -e "$d_test/lib/gdc-dg.exp" && rm "$d_test/lib/gdc-dg.exp"
+if test -e "$d_test/gdc.test" -o -e "$d_test/lib/gdc.exp" -o -e "$d_test/lib/gdc-dg.exp"; then
+    echo "error: cannot update gcc source, please remove D testsuite sources by hand."
+    exit 1
+fi
+
 
 # 1. Copy sources
-ln -s "$top/gcc/d" "$d_gccsrc/gcc/d" && \
-  mkdir "$d_gccsrc/libphobos"        && \
-  cd "$d_gccsrc/libphobos"           && \
-  ../symlink-tree "$top/libphobos"   && \
+ln -s "$top/gcc/d" "$d_gccsrc/gcc/d"   && \
+  mkdir "$d_gccsrc/libphobos"          && \
+  cd "$d_gccsrc/libphobos"             && \
+  ../symlink-tree "$top/libphobos"     && \
+  cd "../gcc/testsuite"                && \
+  ../../symlink-tree "$top/gcc/testsuite" && \
   cd $top
